# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AI Resume Review System API
  description: |
    REST API for the AI Resume Review System that allows users to upload resumes,
    get AI-powered analysis, and manage their review history with secure authentication.
  version: 1.0.0
  contact:
    name: AI Resume Review Team
    email: support@resumereview.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://api-dev.resumereview.com/api/v1
    description: Development environment
  - url: https://api.resumereview.com/api/v1
    description: Production environment
security:
  - BearerAuth: []
paths:
...
  /upload/resume:
    post:
      tags:
        - File Upload
      summary: Upload resume for analysis
      description: |
        Upload a resume file (PDF, DOC, or DOCX) for AI-powered analysis.
        
        **Features:**
        - Supports PDF, DOC, and DOCX files
        - Maximum file size: 30MB
        - Comprehensive validation (file type, size, content)
        - Rate limited: 30 uploads per hour
        - Secure UUID-based file storage
        - Integration with analysis pipeline
        
        **Process:**
        1. Rate limit check (30 uploads/hour)
        2. File validation (type, size, structure)
        3. Secure storage with UUID naming
        4. Database record creation
        5. Queue for AI analysis
      operationId: uploadResume
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Resume file (PDF, DOC, or DOCX, max 30MB)
                target_role:
                  type: string
                  maxLength: 255
                  description: Target job role for analysis
                  example: "Software Engineer"
                target_industry:
                  type: string
                  maxLength: 255
                  description: Target industry for analysis
                  example: "Technology"
                experience_level:
                  type: string
                  enum: ["entry", "mid", "senior", "executive"]
                  description: Experience level for tailored analysis
                  example: "mid"
      responses:
        '201':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileUploadResponse'
        '400':
          description: File validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileUploadError'
              examples:
                validation_failed:
                  summary: File validation failed
                  value:
                    error: "FILE_VALIDATION_FAILED"
                    message: "File validation failed"
                    validation_errors: ["File type 'text/plain' not allowed. Allowed: application/pdf, application/msword, application/vnd.openxmlformats-officedocument.wordprocessingml.document"]
                empty_file:
                  summary: Empty file
                  value:
                    detail: "File is empty"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '413':
          description: File too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "File size (35.2MB) exceeds maximum allowed size (30.0MB)"
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "Too many requests for file_upload"
                block_time_remaining: 1800
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /upload/list:
    get:
      tags:
        - File Upload
      summary: List user's uploaded files
      description: |
        Get a paginated list of files uploaded by the current user.
        Results are ordered by upload date (newest first).
      operationId: listUploads
      parameters:
        - name: page
          in: query
          description: Page number (1-based)
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            example: 1
        - name: per_page
          in: query
          description: Items per page (1-100)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            example: 20
      responses:
        '200':
          description: List of uploads retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /upload/{upload_id}/status:
    get:
      tags:
        - File Upload
      summary: Get upload status
      description: |
        Get the current processing status of a specific upload.
        Shows progress through the analysis pipeline.
      operationId: getUploadStatus
      parameters:
        - name: upload_id
          in: path
          required: true
          description: Upload ID
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Upload status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadStatusResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Upload not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "Upload not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /upload/{upload_id}:
    delete:
      tags:
        - File Upload
      summary: Delete upload
      description: |
        Delete an uploaded file and its associated analysis data.
        This action cannot be undone.
      operationId: deleteUpload
      parameters:
        - name: upload_id
          in: path
          required: true
          description: Upload ID to delete
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Upload deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadDeleteResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Upload not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "Upload not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /upload/stats:
    get:
      tags:
        - File Upload
      summary: Get upload statistics
      description: |
        Get statistics about the current user's uploads including
        total counts, file types, and rate limit status.
      operationId: getUploadStats
      responses:
        '200':
          description: Upload statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadStatsResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from /auth/login endpoint.
        Format: "Bearer <jwt_token>"
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message describing what went wrong
          example: "Invalid email or password"
        code:
          type: string
          description: Optional error code for programmatic handling
          example: "AUTH_INVALID_CREDENTIALS"
        errors:
          type: array
          description: Optional list of validation errors
          items:
            type: object
            properties:
              field:
                type: string
                description: Field that caused the error
                example: "email"
              message:
                type: string
                description: Specific error message
                example: "Invalid email format"

    # File Upload Schemas
    FileUploadResponse:
      type: object
      required:
        - id
        - original_filename
        - file_size
        - mime_type
        - status
        - created_at
        - validation_info
      properties:
        id:
          type: string
          format: uuid
          description: Unique analysis request ID
          example: "550e8400-e29b-41d4-a716-446655440000"
        original_filename:
          type: string
          description: Original filename
          example: "resume.pdf"
        file_size:
          type: integer
          description: File size in bytes
          example: 1048576
        mime_type:
          type: string
          description: File MIME type
          example: "application/pdf"
        status:
          type: string
          enum: ["pending", "processing", "completed", "failed"]
          description: Current processing status
          example: "pending"
        target_role:
          type: string
          description: Target job role
          example: "Software Engineer"
        target_industry:
          type: string
          description: Target industry
          example: "Technology"
        experience_level:
          type: string
          enum: ["entry", "mid", "senior", "executive"]
          description: Experience level
          example: "mid"
        created_at:
          type: string
          format: date-time
          description: Upload timestamp
          example: "2024-01-15T10:30:00Z"
        validation_info:
          $ref: '#/components/schemas/FileValidationInfo'
...
    FileUploadError:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type
          example: "FILE_VALIDATION_FAILED"
        message:
          type: string
          description: Human-readable error message
          example: "File validation failed"
        details:
          type: object
          description: Additional error details
          additionalProperties: true
        validation_errors:
          type: array
          description: File validation errors
          items:
            type: string
          example: ["File type 'text/plain' not allowed"]
    UploadListResponse:
      type: object
      required:
        - uploads
        - total
        - page
        - per_page
        - has_more
      properties:
        uploads:
          type: array
          description: List of user uploads
          items:
            $ref: '#/components/schemas/FileUploadResponse'
        total:
          type: integer
          description: Total number of uploads
          example: 15
        page:
          type: integer
          description: Current page number
          example: 1
        per_page:
          type: integer
          description: Items per page
          example: 20
        has_more:
          type: boolean
          description: Whether more pages exist
          example: false
    UploadStatusResponse:
      type: object
      required:
        - id
        - status
        - original_filename
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Analysis request ID
          example: "550e8400-e29b-41d4-a716-446655440000"
        status:
          type: string
          enum: ["pending", "processing", "completed", "failed"]
          description: Current status
          example: "processing"
        original_filename:
          type: string
          description: Original filename
          example: "resume.pdf"
        created_at:
          type: string
          format: date-time
          description: Upload timestamp
          example: "2024-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2024-01-15T10:35:00Z"
        completed_at:
          type: string
          format: date-time
          description: Completion timestamp
          example: "2024-01-15T10:40:00Z"
        error_message:
          type: string
          description: Error message if failed
          example: "Analysis pipeline error"
        progress_info:
          type: object
          description: Processing progress info
          additionalProperties: true
          example:
            step: "text_extraction"
            progress: 75
    UploadDeleteResponse:
      type: object
      required:
        - success
        - message
        - id
      properties:
        success:
          type: boolean
          description: Whether deletion was successful
          example: true
        message:
          type: string
          description: Success or error message
          example: "Upload 'resume.pdf' deleted successfully"
        id:
          type: string
          format: uuid
          description: ID of deleted upload
          example: "550e8400-e29b-41d4-a716-446655440000"
    UploadStatsResponse:
      type: object
      required:
        - stats
        - rate_limit_info
      properties:
        stats:
          $ref: '#/components/schemas/UploadStats'
        rate_limit_info:
          type: object
          description: Current rate limit status
          properties:
            uploads_allowed_per_hour:
              type: integer
              example: 30
            uploads_remaining:
              type: integer
              example: 25
            reset_time:
              type: string
              format: date-time
              example: "2024-01-15T11:00:00Z"
...
  responses:
    UnauthorizedError:
      description: Invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Could not validate credentials"
...
    ValidationError:
      description: Request validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Validation error"
            errors:
              - field: "email"
                message: "Invalid email format"
              - field: "password"
                message: "Password must be at least 8 characters"
...